requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
twilio
python-dotenv
